Introduced support for serialized parse tables for performance reasons. When Stratego.tbl is requested SGLR looks for Stratego.tbl.bin in the same folder. If not found, the original table is loaded from BAF format and then serialized to the *tbl.bin file.
Unsure if the loaded table is identical with the original one; we attempt to satisfy maximal sharing by overriding readResolve(). Unsure if this is sufficient. Risk: the match at the end of each unit might fail for the wrong reason. When in doubt remove the *.tbl.bin file and try again.

1) Loading of parse table from serialized form: SGLR
See all ATerm*Impl classes for readObject() implementation which controls what instances of ATerm are loaded by the deserialization; Same technique like Jakarta Enum.
2) Avoid parsing: Amb, Node
3) Dump the parse tree in case it does not match: ParseTestCase

svn path=/spoofax/trunk/spoofax/org.spoofax.jsglr/; revision=14952