Fix a parser bug.
When a String is being parsed and the parse step at the last character of the string causes a lookahead reduction,
the PushbackStringIterator's position gets mangled, causing that last character to be read (and parsed) twice.

This bug was discovered by a compiler construction test case:
```
language MiniJava
start symbol ID

test no slash in ID [[a/]] parse fails
```

Because of the bug, this got parsed as `a//` instead of `a/`.
As `//` is the start of a comment, which is layout, this parsed correctly to an ID.