Make only one call to parse in integrationtest
The test for success used to work as follows:
- Call only Parser.parse
- Assert that parse is successful
- Call JSGLR2.parse (which also calls Parser.parse)
- Assert that imploding is successful

I've changed the test to call parseUnsafe instead.
This will throw an exception in the case that parsing fails,
and return `null` in the case that imploding fails.
This means that we can still distinguish failures between the two stages
but only need to call the Parser.parse method once.