Remove state field from IncrementalParseForest
The state field is only necessary for `IncrementalDerivation`s.
Wagner's algorithm defines symbol nodes (ParseNode in JSGLR2) as having
multiple states by default. If a symbol node only has one derivation,
the symbol node is left out by Wagner. This means that the state check
in the shifter (`getCharacterNodeToShift`), which already only happens
if there is only one state, can be done with the state of the first
derivation of the parse node instead.

While this simplifies some code, the IncrementalParseForestManager still
needs to know the current state when creating a Derivation node.
Currently, I pass this via the Parse object, but this is damn ugly.
In an ideal world, I would just directly pass the state to the IPFM,
like I would pass it to the constructor of the Derivation node.
(added a TODO for this)