Decouple Parse state from Imploder state
It used to be the case that the Parse (AbstractParse) state contained a
reference to a Tokens object, that would be passed along during
imploding.
Now, the Tokens object is only created inside the Imploder, as the
parser does not need to know anything about this Tokens object.
To do this, the IImploder interface has been updated to:
    implode(String input, String filename, ParseForest parseForest)

Related to this, also the Result classes have been decoupled.
ParseResult and ImplodeResult now only have one type parameter*,
as the first only needs to know the type of the ParseForest,
while the second only needs to know the type of the AbstractSyntaxTree.
The class JSGLR2Result is now constructed as the combination of a
ParseResult and an ImplodeResult, rather than the ParseResult being
threaded through to the ImplodeResult.

*) This saves on a whole bunch of `?` in type declarations :smile: