Pass StackNode to ParseForestManager instead of just Position
This will be useful for incremental parsing, as the state of the stack
will be saved in the parse nodes.
The LayoutSensitiveParseForestManager can still use the stack position
as `beginPosition`, as it used to.
Other ParseForestManagers are not affected (besides method signature
changes), as they do not use the Position anymore.