Reduce duplicate code in incremental imploder
Instead of adding an extra parameter to implodeParseNode and
implodeDerivation, the first input parameter is now a generic object
that can contain whichever input needs to be passed to these methods.

For the regular imploder, this object will only contain the input
string. For the incremental imploder, this object will also contain the
WeakHashMap that is used as cache.

To properly override implodeParseNode in IncrementalTreeImploder, the
type of the input object needs to be a type parameter of the class.
Because of this, there needs to be a factory that creates the input
object.