Rename IIncrementalParseState.isMultipleStates to newParseNodesAreReusable
This should make the meaning of the method more clear.
In addition to this, the `state` field has been moved from
IncrementalDerivation to IncrementalParseNode, which removes a layer of
indirection in most cases.
The `state` field is set to `NO_STATE` when a new derivation is added,
following Wagner's algorithm more closely.
To make the algorithm even more understandable, IncrementalParseNodes
now have a method named `isReusable` and no longer expose the `state`
field directly.