Fix imploder attachments for injections on empty trees
In the added test for an empty list, the separated imploders/tokenizers
would throw NullPointerExceptions in configureInjection because there
would be null tokens on their subtrees, or in other words, there would
be no ImploderAttachment to configure the injection for.

This commit makes sure that every context-free subtree has tokens.
For empty subtrees, the `parentLeftToken` is used.